src/main/java/ch02/Pizza.java
package ch02;

import java.util.Collections;
import java.util.EnumSet;
import java.util.Objects;
import java.util.Set;

/**
 * The Pizza class demonstrates the Builder pattern for class hierarchies
 * Effective Java, 3rd Edition
 * Item 2: Consider a builder when faced with many constructor parameters
 */
public abstract class Pizza {
  public enum Topping {HAM, MUSHROOM, ONION, PEPPER, SAUSAGE}

  final Set<Topping> toppings;

  abstract static class Builder<T extends Builder<T>> {
    EnumSet<Topping> toppings = EnumSet.noneOf(Topping.class);

    public T addTopping(Topping topping) {
      toppings.add(Objects.requireNonNull(topping));
      return self();
    }

    abstract Pizza build();

    // Subclasses must override this method to return "this"
    protected abstract T self();
  }

  Pizza(Builder<?> builder) {
    toppings = builder.toppings.clone(); // See Item 50
  }

  public Set<Topping> getToppings() {
    return Collections.unmodifiableSet(toppings);
  }
}